feat: OAuth testing with mock_auth_hash
