+++ app/controllers/welcome_controller.rb
@@ -1,4 +1,7 @@
 class WelcomeController < ApplicationController
   def index
   end
+  def error
+    @message = params[:message]
+  end
 end
