+++ app/views/users/new.html.erb
@@ -1,0 +1,2 @@
+<h1>Create your account</h1>
+<%= button_to "Join with Google", '/auth/google_oauth2', method: :get %>
