+++ app/views/users/show.html.erb
@@ -1,3 +1,4 @@
 <h1>You are Logged-In!</h1>
 
 <%= @user.name %>
+<%= button_to 'Sign out', root_path %>
