+++ app/views/welcome/error.html.erb
@@ -1,0 +1,1 @@
+<%= @message %>
