+++ app/views/welcome/index.html.erb
@@ -1,1 +1,3 @@
 <h1>Get Yo Newz Hurr</h1>
+<%= link_to "Create an account", '/register', method: :get %>
+<%= link_to "Sign in with Google", '/auth/google_oauth2', method: :get %>
