+++ spec/spec_helper.rb
@@ -1,4 +1,6 @@
 require 'simplecov'
+require 'omniauth'
+require 'capybara/rspec'
 SimpleCov.start
 
 # This file was generated by the `rails generate rspec:install` command. Conventionally, all
@@ -94,5 +96,6 @@ RSpec.configure do |config|
   # as the one that triggered the failure.
   Kernel.srand config.seed
 =end
+
 end
 OmniAuth.config.test_mode = true
